Adjusted migration script

+++ 5/debian-10/rootfs/tfc_wp-init.sh
@@ -17,6 +17,8 @@ LATESTVERSION="$PERSPATH/.lastversioninstalled";
 WPTEMPDIR="$VOLPATH/tmp";
 # WP Content Dir Location
 WPCONTENTDIR="$PERSPATH/wp-content";
+# Migration flag (overritten on an actual migration)
+MIGRATIONFLAG="no";
 # Stop defining variables
 
 # This placeholder function should run only when the container is executed as root so we can migrate to non-root environment
@@ -34,6 +36,8 @@ if [ -d "/bitnami/wordpress" ]; then
 
   info "We should now rename that old wordpress install so we don't re-import it again in the future"
   mv /bitnami/wordpress /bitnami/bitnami_wp_backup
+
+  MIGRATIONFLAG="yes";
 
 fi
 # This placeholder function should run only when the container is executed as root so we can migrate to non-root environment
@@ -108,6 +112,9 @@ PHP
   info "Creating placeholder files"
   touch "$VOLPATH"/.htaccess
 
+# Are we forcing a fresh start?
+if [ $MIGRATIONFLAG = "no" ]; then
+
   info "Cleanup the placeholder files before moving them"
   rm -rf "$PERSPATH"/.htaccess "$PERSPATH"/wp-config.php "$WPCONTENTDIR"
 
@@ -115,6 +122,15 @@ PHP
   mv -f "$VOLPATH"/.htaccess "$PERSPATH"/.htaccess
   mv -f "$VOLPATH"/wp-config.php "$PERSPATH"/wp-config.php
   mv -f "$VOLPATH"/wp-content "$WPCONTENTDIR"
+
+elif [ $MIGRATIONFLAG = "yes" ]; then
+
+  info "Remove virgin install default files from volatile storage"
+  rm -rf "$VOLPATH"/.htaccess
+  rm -rf "$VOLPATH"/wp-config.php
+  rm -rf "$VOLPATH"/wp-content
+
+fi
 
   info "Setup placeholder install file"
   touch "$INSTALLFILE"
